basic_math: NOT, AND, OR, EOR: do not cast operands to unsigned 32-bit
Doing so breaks the use as logical operators.

This was done to fix bitwise operations on 32-bit color values, which
still works when casting to int64_t instead.

When operating on unsigned 64-bit values (such as 64-bit pointers) this
will still fail, but that can probably only be fixed by having dedicated
bitwise operators.

Fixes HB1/AND_3.BAS.

+++ ttbasic/basic_math.cpp
@@ -373,7 +373,7 @@ Like most BASIC implementations, Engine BASIC does not have a dedicated
 "logical NOT" operator; instead, the bitwise operator is used.
 ***/
   case I_BITREV: // NOT
-    return ~((uint32_t)(int64_t)irel());
+    return ~((int64_t)irel());
   default:
     cip--;
     value = irel();
@@ -395,7 +395,7 @@ Like most BASIC implementations, Engine BASIC does not have a dedicated
 "logical AND" operator; instead, the bitwise operator is used.
 ***/
       tmp = iand();
-      value = ((uint32_t)(int64_t)value) & ((uint32_t)(int64_t)tmp);
+      value = ((int64_t)value) & ((int64_t)tmp);
       break;
     default:
       cip--;
@@ -424,7 +424,7 @@ Like most BASIC implementations, Engine BASIC does not have a dedicated
 ***/
     case I_OR:
       tmp = iand();
-      value = ((uint32_t)(int64_t)value) | ((uint32_t)(int64_t)tmp);
+      value = ((int64_t)value) | ((int64_t)tmp);
       break;
 /***bo op EOR
 Bitwise exclusive-OR operator.
@@ -434,7 +434,7 @@ Bitwise exclusive-OR operator.
 ***/
     case I_XOR:
       tmp = iand();
-      value = ((uint32_t)(int64_t)value) ^ ((uint32_t)(int64_t)tmp);
+      value = ((int64_t)value) ^ ((int64_t)tmp);
       break;
     default:
       cip--;
